+++ game-engine/stratego-lib/src/parse.rs
@@ -22,17 +22,18 @@ fn parse_str_to_i16(s: &str) -> i16 {
     s.parse::<i16>().unwrap()
 }
 
-#[test]
-fn should_parse_input() {
-    assert_eq!((0, 0), parse_input("0A"));
-    assert_eq!((4, 1), parse_input("4B"));
-    assert_eq!((3, 7), parse_input("3H"));
-}
 
 pub fn parse_i16_to_str(x: i16) -> String {
     let x = x.to_be_bytes();
     String::from_utf8(x.to_vec())
         .unwrap()
+}
+
+#[test]
+fn should_parse_input() {
+    assert_eq!((0, 0), parse_input("0A"));
+    assert_eq!((4, 1), parse_input("4B"));
+    assert_eq!((3, 7), parse_input("3H"));
 }
 
 #[test]
